python-service/requirements.txt
rembg
flask
pillow